Refactor to use fragments
