+++ app/src/main/java/com/snijsure/omdbsearch/di/modules/UiModule.kt
@@ -2,6 +2,7 @@ package com.snijsure.omdbsearch.di.modules
 
 import com.snijsure.omdbsearch.ui.view.MovieDetailActivity
 import com.snijsure.omdbsearch.ui.view.MovieListActivity
+import com.snijsure.omdbsearch.ui.view.MovieListFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
@@ -13,4 +14,6 @@ abstract class UiModule {
 
     @ContributesAndroidInjector
     abstract fun contributeMovieListActivity(): MovieListActivity
+    @ContributesAndroidInjector
+    abstract fun contributeMovieListFragment(): MovieListFragment
 }
