+++ app/src/main/res/layout/activity_movie_list.xml
@@ -1,85 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                                             xmlns:app="http://schemas.android.com/apk/res-auto"
-                                             xmlns:tools="http://schemas.android.com/tools"
-                                             android:id="@+id/rootView"
-                                             android:layout_width="match_parent"
-                                             android:layout_height="match_parent"
-                                             android:background="@android:color/darker_gray"
-                                             tools:context=".ui.view.MovieListActivity">
-
-
-    <android.support.v7.widget.AppCompatTextView
-            android:id="@+id/no_connection"
-            android:layout_width="140dp"
-            android:layout_height="140dp"
-            android:visibility="gone"
-            android:contentDescription="@string/no_connection"
-            android:text="@string/no_connection"
-            android:drawableTop="@drawable/ic_no_network"
-            android:textSize="18sp"
-            android:textStyle="bold"
-            android:gravity="center"
-            app:layout_constraintCircle="@id/rootView"
-            app:layout_constraintCircleRadius="0dp"
-            android:layout_margin="24dp"/>
-
-    <android.support.v7.widget.AppCompatTextView
-            android:id="@+id/welcome_text"
-            android:layout_width="200dp"
-            android:layout_height="140dp"
-            android:visibility="visible"
-            android:contentDescription="@string/welcome_text"
-            android:text="@string/welcome_text"
-            android:drawableTop="@drawable/ic_outline_movie_24px"
-            android:textAppearance="@android:style/TextAppearance.Holo.Large"
-            android:textStyle="bold"
-            android:gravity="center"
-            app:layout_constraintCircle="@id/rootView"
-            app:layout_constraintCircleRadius="0dp"
-    />
-
-
-    <RelativeLayout android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:id="@+id/busy_indicator"
-                    android:orientation="vertical"
-                    android:visibility="gone">
-        <ProgressBar
-                android:id="@+id/progressbar"
-                android:layout_width="40dp"
-                android:layout_height="40dp"
-                android:orientation="vertical"
-                android:gravity="center"
-                style="@android:style/Widget.ProgressBar.Large.Inverse"
-                android:layout_marginEnd="24dp"
-                android:layout_marginStart="24dp"
-                android:layout_marginTop="24dp"
-                android:layout_marginBottom="24dp"
-                android:layout_centerInParent="true"
-        />
-        <android.support.v7.widget.AppCompatTextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_centerHorizontal="true"
-                android:textStyle="bold"
-                android:textSize="16sp"
-                android:maxLines="3"
-                android:padding="8dp"
-                android:layout_marginTop="16dp"
-                android:text="@string/loading_text"
-                android:layout_below="@id/progressbar"/>
-    </RelativeLayout>
-
-    <android.support.v7.widget.RecyclerView
-            android:id="@+id/movie_list_view"
-            android:visibility="gone"
+<android.support.design.widget.CoordinatorLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        android:id="@+id/drawer_layout"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+    <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:padding="4dp"
-            android:scrollbars="vertical"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"/>
-</android.support.constraint.ConstraintLayout>
+            android:orientation="vertical">
+        <FrameLayout
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:id="@+id/my_nav_host_fragment"
+        />
+    </LinearLayout>
+</android.support.design.widget.CoordinatorLayout>
